requirements.txt
pyswisseph
reportlab
geopy
timezonefinder
tzdata
flask
